+++ src/tests/rtt/rtt.cpp
@@ -50,7 +50,7 @@ public:
 
     virtual bool startup( const unsigned int numContexts );
     virtual bool init();
-    virtual bool frame( const gmtl::Matrix44f& view, const gmtl::Matrix44f& proj );
+    virtual bool frame( const gmtl::Matrix44d& view, const gmtl::Matrix44d& proj );
     virtual bool shutdown()
     {
         return( true );
@@ -312,7 +312,7 @@ bool RttDemo::init()
     return( true );
 }
 
-bool RttDemo::frame( const gmtl::Matrix44f& view, const gmtl::Matrix44f& proj )
+bool RttDemo::frame( const gmtl::Matrix44d& view, const gmtl::Matrix44d& proj )
 {
     if( !getStartupCalled() )
         return( true );
